requirements.txt
# OCR处理流程
torch>=2.0.1
torchvision>=0.15.2
transformers>=4.37.2
tiktoken>=0.6.0
verovio>=5.3.0
accelerate>=1.7.0




# 基础依赖
torch>=1.10.0
transformers>=4.18.0
numpy>=1.20.0
pandas>=1.3.0
tqdm>=4.62.0
pillow>=8.3.0
matplotlib>=3.4.0
modelscope==1.8.3
datasets==2.14.7

# OCR相关
paddlepaddle-gpu>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.5.0
scikit-image>=0.19.0
imutils>=0.5.4

# 文本处理
jieba>=0.42.0
nltk>=3.6.0

# Web服务（如果需要）
fastapi>=0.70.0
uvicorn>=0.15.0